requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.9.0
beautifulsoup4==4.13.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cobble==0.1.4
cssselect2==0.8.0
distlib==0.3.9
distro==1.9.0
EbookLib==0.18
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.1
fsspec==2025.5.1
ftfy==6.3.1
google-auth==2.40.2
google-genai==1.17.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
identify==2.6.12
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lxml==5.4.0
mammoth==1.9.1
markdown2==2.5.3
markdownify==0.13.1
marker-pdf==1.7.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ollama==0.5.1
openai==1.82.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pdftext==0.6.2
pillow==11.2.1
platformdirs==4.3.8
pre_commit==4.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydyf==0.11.0
pypdfium2==4.30.0
pyphen==0.17.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
surya-ocr==0.14.3
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
virtualenv==20.31.2
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
websockets==15.0.1
XlsxWriter==3.2.3
zopfli==0.2.3.post1
